move models to shared project
