--- Server/Data/Models/Entity.cs
@@ -1,11 +1,0 @@
-namespace BlazorCMS.Server.Data.Models
-{
-    public class Entity
-    {
-        #region Properties
-
-        public long Id { get; set; }
-
-        #endregion Properties
-    }
-}
