+++ Shared/Models/Entity.cs
@@ -1,0 +1,11 @@
+namespace BlazorCMS.Shared.Models
+{
+    public class Entity
+    {
+        #region Properties
+
+        public long Id { get; set; }
+
+        #endregion Properties
+    }
+}
