+++ Shared/Models/Section.cs
@@ -1,0 +1,19 @@
+using System.Collections.Generic;
+
+namespace BlazorCMS.Shared.Models
+{
+    public class Section : Entity
+    {
+        #region Properties
+
+        public string Name { get; set; }
+
+        #endregion Properties
+
+        #region Navigation Properties
+
+        public IEnumerable<Article> Articles { get; set; }
+
+        #endregion Navigation Properties
+    }
+}
